scripts/requirements.txt
firebase-admin>=6.0.0
pycryptodome>=3.12.0
